Refactor to shift to "imreg" package.
