+++ examples/linreg.py
@@ -7,12 +7,12 @@ Estimates a linear warp field, the target is a transformed version of lenna:
 import scipy.ndimage as nd
 import scipy.misc as misc
 
-from register.models import model
-from register.metrics import metric
-from register.samplers import sampler
-from register import register
+from imreg.models import model
+from imreg.metrics import metric
+from imreg.samplers import sampler
+from imreg import register
 
-from register.visualize import plot
+from imreg.visualize import plot
 
 # Form some test data (lena, lena rotated 20 degrees)
 image = misc.lena()
