+++ examples/nonlinreg.py
@@ -7,12 +7,12 @@ import numpy as np
 import scipy.ndimage as nd
 import scipy.misc as misc
 
-from register.models import model
-from register.metrics import metric
-from register.samplers import sampler
+from imreg.models import model
+from imreg.metrics import metric
+from imreg.samplers import sampler
 
-from register.visualize import plot
-from register import register
+from imreg.visualize import plot
+from imreg import register
 
 def warp(image):
     """
