+++ imreg/features/haar2d.py
