+++ imreg/metrics/metric.py
@@ -5,13 +5,13 @@ import numpy as np
 class Metric(object):
     """
     Abstract similarity metric.
-    
+
     Attributes
     ----------
     METRIC : string
         The type of similarity metric being used.
     DESCRIPTION : string
-        A meaningful description of the metric used, with references where 
+        A meaningful description of the metric used, with references where
         appropriate.
     """
 
@@ -24,26 +24,26 @@ class Metric(object):
     def error(self, warpedImage, template):
         """
         Evaluates the metric.
-        
+
         Parameters
         ----------
         warpedImage: nd-array
             Input image after warping.
         template: nd-array
             Template image.
-    
+
         Returns
         -------
         error: nd-array
            Metric evaluated over all image coordinates.
         """
-        
+
         raise NotImplementedError('')
-        
+
     def jacobian(self, model, warpedImage, p=None):
         """
         Computes the jacobian dP/dE.
-        
+
         Parameters
         ----------
         model: deformation model
@@ -52,14 +52,14 @@ class Metric(object):
             Input image after warping.
         p : optional list
             Current warp parameters
-    
+
         Returns
         -------
         jacobian: nd-array
            A derivative of model parameters with respect to the metric.
         """
         raise NotImplementedError('')
-        
+
     def __str__(self):
         return 'Metric: {0} \n {1}'.format(
             self.METRIC,
@@ -86,7 +86,7 @@ class Residual(Metric):
     def jacobian(self, model, warpedImage, p=None):
         """
         Computes the jacobian dP/dE.
-        
+
         Parameters
         ----------
         model: deformation model
@@ -95,7 +95,7 @@ class Residual(Metric):
             Input image after warping.
         p : optional list
             Current warp parameters
-    
+
         Returns
         -------
         jacobian: nd-array
@@ -113,20 +113,20 @@ class Residual(Metric):
 
         J = np.zeros_like(dPx)
         for index in range(0, dPx.shape[1]):
-            J[:,index] = dPx[:,index]*dIx + dPy[:,index]*dIy
+            J[:, index] = dPx[:, index] * dIx + dPy[:, index] * dIy
         return J
 
     def error(self, warpedImage, template):
         """
         Evaluates the residual metric.
-        
+
         Parameters
         ----------
         warpedImage: nd-array
             Input image after warping.
         template: nd-array
             Template image.
-    
+
         Returns
         -------
         error: nd-array
