+++ imreg/register.py
@@ -6,11 +6,11 @@ import scipy.ndimage as nd
 def _smooth(image, variance):
     """
     Gaussian smoothing using the fast-fourier-transform (FFT)
-    
+
     Parameters
     ----------
     image: nd-array
-        Input image 
+        Input image
     variance: float
         Variance of the Gaussian kernel.
 
@@ -23,7 +23,7 @@ def _smooth(image, variance):
     --------
     regisger.Register.smooth
     """
-    
+
     return np.real(
         np.fft.ifft2(
             nd.fourier_gaussian(
@@ -37,7 +37,7 @@ def _smooth(image, variance):
 class Coordinates(object):
     """
     Container for grid coordinates.
-    
+
     Attributes
     ----------
     domain : nd-array
@@ -47,13 +47,13 @@ class Coordinates(object):
     homogenous : nd-array
         `Homogenous` coordinate system representation of grid coordinates.
     """
-    
+
     def __init__(self, domain, spacing=None):
-        
+
         self.spacing = 1.0 if not spacing else spacing
         self.domain = domain
         self.tensor = np.mgrid[0.:domain[1], 0.:domain[3]]
-        
+
         self.homogenous = np.zeros((3,self.tensor[0].size))
         self.homogenous[0] = self.tensor[1].flatten()
         self.homogenous[1] = self.tensor[0].flatten()
@@ -63,7 +63,7 @@ class Coordinates(object):
 class RegisterData(object):
     """
     Container for registration data.
-    
+
     Attributes
     ----------
     data : nd-array
@@ -73,11 +73,11 @@ class RegisterData(object):
     features : dictionary, optional
         A mapping of unique ids to registration features.
     """
-    
+
     def __init__(self, data, coords=None, features=None, spacing=1.0):
 
         self.data = data.astype(np.double)
-        
+
         if not coords:
             self.coords = Coordinates(
                 [0, data.shape[0], 0, data.shape[1]],
@@ -85,47 +85,47 @@ class RegisterData(object):
                 )
         else:
             self.coords = coords
-        
+
         # Features are (as a starting point a dictionary) which define
-        # labeled salient image coordinates (point features). 
-        
+        # labeled salient image coordinates (point features).
+
         self.features = features
-        
-    
+
+
     def downsample(self, factor=2):
         """
-        Down samples the RegisterData by a user defined factor. The ndimage 
-        zoom function is used to interpolate the image, with a scale defined 
+        Down samples the RegisterData by a user defined factor. The ndimage
+        zoom function is used to interpolate the image, with a scale defined
         as 1/factor.
-        
+
         Spacing is used to infer the scale difference between images - defining
         the size of a pixel in arbitrary units (atm).
-        
+
         Parameters
         ----------
         factor: float (optional)
             The scaling factor which is applied to image data and coordinates.
-            
+
         Returns
         -------
         scaled: nd-array
            The parameter update vector.
         """
-        
+
         resampled = nd.zoom(self.data, 1. / factor)
-        
+
         # TODO: features need to be scaled also.
         return RegisterData(resampled, spacing=factor)
-    
+
     def smooth(self, variance):
         """
         Smooth feature data in place.
-    
+
         Parameters
         ----------
         variance: float
             Variance of the Gaussian kernel.
-       
+
         See also
         --------
         register.Register.smooth
@@ -137,7 +137,7 @@ class RegisterData(object):
 class optStep():
     """
     A container class for optimization steps.
-    
+
     Attributes
     ----------
     warpedImage: nd-array
@@ -155,21 +155,21 @@ class optStep():
     decreasing: boolean.
         State of the error function at this point.
     """
-    
+
     def __init__(
-        self, 
+        self,
         warpedImage=None,
         warp=None,
-        grid=None, 
-        error=None, 
+        grid=None,
+        error=None,
         p=None,
-        deltaP=None, 
+        deltaP=None,
         decreasing=None,
         template=None,
         image=None,
         displacement=None
         ):
-        
+
         self.warpedImage = warpedImage
         self.warp = warp
         self.grid = grid
@@ -186,7 +186,7 @@ class Register(object):
     """
     A registration class for estimating the deformation model parameters that
     best solve:
-    
+
     | :math:`f( W(I;p), T )`
     |
     | where:
@@ -194,15 +194,15 @@ class Register(object):
     |    :math:`W(x;p)`: is a deformation model (defined by the parameter set p).
     |    :math:`I`     : is an input image (to be deformed).
     |    :math:`T`     : is a template (which is a deformed version of the input).
-    
+
     Notes:
     ------
-    
+
     Solved using a modified gradient descent algorithm.
-    
-    .. [0] Levernberg-Marquardt algorithm, 
-           http://en.wikipedia.org/wiki/Levenberg-Marquardt_algorithm 
-    
+
+    .. [0] Levernberg-Marquardt algorithm,
+           http://en.wikipedia.org/wiki/Levenberg-Marquardt_algorithm
+
     Attributes
     ----------
     model: class
@@ -212,12 +212,12 @@ class Register(object):
     sampler: class
         A `sampler` class definition.
     """
-    
-   
-    
+
+
+
     MAX_ITER = 200
     MAX_BAD = 5
-    
+
     def __init__(self, model, metric, sampler):
 
         self.model = model
@@ -227,18 +227,18 @@ class Register(object):
     def __deltaP(self, J, e, alpha, p=None):
         """
         Computes the parameter update.
-        
+
         Parameters
         ----------
         J: nd-array
             The (dE/dP) the relationship between image differences and model
             parameters.
         e: float
-            The evaluated similarity metric.    
+            The evaluated similarity metric.
         alpha: float
             A dampening factor.
         p: nd-array or list of floats, optional
-        
+
         Returns
         -------
         deltaP: nd-array
@@ -254,14 +254,14 @@ class Register(object):
     def __dampening(self, alpha, decreasing):
         """
         Computes the adjusted dampening factor.
-        
+
         Parameters
         ----------
         alpha: float
             The current dampening factor.
         decreasing: boolean
             Conditional on the decreasing error function.
-            
+
         Returns
         -------
         alpha: float
@@ -279,7 +279,7 @@ class Register(object):
                  verbose=False):
         """
         Computes the registration between the image and template.
-        
+
         Parameters
         ----------
         image: nd-array
@@ -295,8 +295,8 @@ class Register(object):
         plotCB: function, optional
             A plotting function.
         verbose: boolean
-            A debug flag for text status updates. 
-        
+            A debug flag for text status updates.
+
         Returns
         -------
         p: nd-array.
@@ -308,59 +308,59 @@ class Register(object):
         error: float
             Fitting error.
         """
-        
+
         #TODO: Determine the common coordinate system.
         if image.coords.spacing != template.coords.spacing:
              raise ValueError('Coordinate systems differ.')
-            
+
         # Initialize the models, metric and sampler.
         model = self.model(image.coords)
         sampler = self.sampler(image.coords)
         metric = self.metric()
-        
+
         if displacement is not None:
-            
+
             # Account for difference warp resolutions.
             scale = (
                 (image.data.shape[0] * 1.) / displacement.shape[1],
                 (image.data.shape[1] * 1.) / displacement.shape[2],
                 )
-            
-            # Scale the displacement field and estimate the model parameters, 
+
+            # Scale the displacement field and estimate the model parameters,
             # refer to test_CubicSpline_estimate
-            scaledDisplacement = np.array([ 
-                nd.zoom(displacement[0], scale), 
+            scaledDisplacement = np.array([
+                nd.zoom(displacement[0], scale),
                 nd.zoom(displacement[1], scale)
                 ]) * scale[0]
-            
+
             # Estimate p, using the displacement field.
             p = model.estimate(-1.0*scaledDisplacement)
-        
+
         p = model.identity if p is None else p
         deltaP = np.zeros_like(p)
 
         # Dampening factor.
         alpha = alpha if alpha is not None else 1e-4
-        
+
         # Variables used to implement a back-tracking algorithm.
         search = []
         badSteps = 0
         bestStep = None
 
         for itteration in range(0,self.MAX_ITER):
-            
-            # Compute the inverse "warp" field. 
+
+            # Compute the inverse "warp" field.
             warp = model.warp(p)
-            
+
             # Sample the image using the inverse warp.
             warpedImage = _smooth(
                 sampler.f(image.data, warp).reshape(image.data.shape),
                 0.50,
                 )
-            
+
             # Evaluate the error metric.
             e = metric.error(warpedImage, template.data)
-            
+
             # Cache the optimization step.
             searchStep = optStep(
                error=np.abs(e).sum()/np.prod(image.data.shape),
@@ -374,10 +374,10 @@ class Register(object):
                image=image.data,
                decreasing=True
                )
-            
+
             # Update the current best step.
             bestStep = searchStep if bestStep is None else bestStep
-            
+
             if verbose:
                 print ('{0}\n'
                        'iteration  : {1} \n'
@@ -395,7 +395,7 @@ class Register(object):
             search.append(searchStep)
 
             if len(search) > 1:
-                
+
                 searchStep.decreasing = (searchStep.error < bestStep.error)
 
                 alpha = self.__dampening(
@@ -404,7 +404,7 @@ class Register(object):
                     )
 
                 if searchStep.decreasing:
-                    
+
                     bestStep = searchStep
 
                     if plotCB is not None:
@@ -412,13 +412,13 @@ class Register(object):
                                template.data,
                                warpedImage,
                                image.coords.tensor,
-                               warp, 
+                               warp,
                                '{0}:{1}'.format(model.MODEL, itteration)
                                )
                 else:
-                    
+
                     badSteps += 1
-                    
+
                     if badSteps > self.MAX_BAD:
                         if verbose:
                             print ('Optimization break, maximum number '
@@ -432,7 +432,7 @@ class Register(object):
             # parameters.
 
             J = metric.jacobian(model, warpedImage, p)
-            
+
             # Compute the parameter update vector.
             deltaP = self.__deltaP(
                 J,
@@ -444,7 +444,7 @@ class Register(object):
             # Evaluate stopping condition:
             if np.dot(deltaP.T, deltaP) < 1e-4:
                 break
-            
+
             # Update the estimated parameters.
             p += deltaP
 
@@ -453,29 +453,29 @@ class Register(object):
 class KybicRegister(Register):
     """
     Variant of LM algorithm as described by:
-      
+
     Kybic, J. and Unser, M. (2003). Fast parametric elastic image
         registration. IEEE Transactions on Image Processing, 12(11), 1427-1442.
     """
-    
+
     def __init__(self, model, metric, sampler):
         Register.__init__(self, model, metric, sampler)
 
     def __deltaP(self, J, e, alpha, p):
         """
         Computes the parameter update.
-        
+
         Parameters
         ----------
         J: nd-array
             The (dE/dP) the relationship between image differences and model
             parameters.
         e: float
-            The evaluated similarity metric.    
+            The evaluated similarity metric.
         alpha: float
             A dampening factor.
         p: nd-array or list of floats, optional
-        
+
         Returns
         -------
         deltaP: nd-array
@@ -491,14 +491,14 @@ class KybicRegister(Register):
     def __dampening(self, alpha, decreasing):
         """
         Computes the adjusted dampening factor.
-        
+
         Parameters
         ----------
         alpha: float
             The current dampening factor.
         decreasing: boolean
             Conditional on the decreasing error function.
-            
+
         Returns
         -------
         alpha: float
@@ -511,19 +511,19 @@ class FeatureRegister():
     """
     A registration class for estimating the deformation model parameters that
     best solve:
-    
+
     | :math:`\arg\min_{p} | f(p_0) - p_1 |`
     |
     | where:
     |    :math:`f`     : is a transformation function.
     |    :math:`p_0`   : image features.
     |    :math:`p_1`   : template features.
-     
+
     Notes:
     ------
-    
+
     Solved using linear algebra - does not consider pixel intensities
-    
+
     Attributes
     ----------
     model: class
@@ -531,7 +531,7 @@ class FeatureRegister():
     sampler: class
         A `sampler` class definition.
     """
-    
+
     def __init__(self, model, sampler):
 
         self.model = model
@@ -540,7 +540,7 @@ class FeatureRegister():
     def register(self, image, template):
         """
         Computes the registration using only image (point) features.
-            
+
         Parameters
         ----------
         image: RegisterData
@@ -549,7 +549,7 @@ class FeatureRegister():
             The target registration data.
         model: Model
             The deformation model.
-            
+
         Returns
         -------
         p: nd-array.
@@ -561,33 +561,33 @@ class FeatureRegister():
         error: float
             Fitting error.
         """
-        
+
         # Initialize the models, metric and sampler.
         model = self.model(image.coords)
         sampler = self.sampler(image.coords)
-        
-        # Form corresponding point sets. 
+
+        # Form corresponding point sets.
         imagePoints = []
         templatePoints = []
-        
+
         for id, point in image.features['points'].items():
             if id in template.features['points']:
                 imagePoints.append(point)
                 templatePoints.append(template.features['points'][id])
                 #print '{} -> {}'.format(imagePoints[-1], templatePoints[-1])
-        
+
         if not imagePoints or not templatePoints:
             raise ValueError('Requires corresponding features to register.')
-        
+
         # Note the inverse warp is estimated here.
         p, error = model.fit(
             np.array(templatePoints),
             np.array(imagePoints)
             )
-        
+
         warp = model.warp(p)
-        
+
         # Sample the image using the inverse warp.
         warpedImage = sampler.f(image.data, warp).reshape(image.data.shape)
-        
+
         return p, warp, warpedImage, error
