+++ imreg/samplers/libsampler.cpp
