+++ imreg/samplers/ndarray.h
