+++ imreg/samplers/numpyctypes.py
