+++ imreg/setup.py
@@ -4,7 +4,7 @@ def configuration(parent_package='', top_path=None):
 
     from numpy.distutils.misc_util import Configuration
 
-    config = Configuration('register', parent_package, top_path)
+    config = Configuration('imreg', parent_package, top_path)
 
     config.add_subpackage('models')
     config.add_subpackage('metrics')
