+++ imreg/visualize/dialog.py
