+++ imreg/visualize/plot.py
