+++ tests/test_register.py
@@ -3,11 +3,11 @@ import numpy as np
 import scipy.ndimage as nd
 import scipy.misc as misc
 
-import register.models.model as model
-import register.metrics.metric as metric
-import register.samplers.sampler as sampler
+import imreg.models.model as model
+import imreg.metrics.metric as metric
+import imreg.samplers.sampler as sampler
 
-from register import register
+from imreg import register
 
 def warp(image, p, model, sampler):
     """
@@ -16,7 +16,7 @@ def warp(image, p, model, sampler):
     coords = register.Coordinates(
         [0, image.shape[0], 0, image.shape[1]]
         )
-    
+
     model = model(coords)
     sampler = sampler(coords)
 
@@ -99,8 +99,8 @@ def test_shift(image, template, p):
     # Coerce the image data into RegisterData.
     image = register.RegisterData(image)
     template = register.RegisterData(template)
-    
-    
+
+
     step, _search = shift.register(
         image,
         template
@@ -127,7 +127,7 @@ def test_affine(image, template, p):
     # Coerce the image data into RegisterData.
     image = register.RegisterData(image)
     template = register.RegisterData(template)
-    
+
     step, _search = affine.register(
         image,
         template
